Fix temporarily the D-Bus property type
The variant for the BIOS integer properties should match the type
of the D-Bus property. This is a temporary fix and should be based
on the property type from the JSON.

+++ libpldmresponder/bios_parser.cpp
@@ -500,7 +500,7 @@ uint64_t getAttrValue(const AttrName& attrName)
         return std::get<AttrDefaultValue>(valueEntry);
     }
 
-    return pldm::utils::DBusHandler().getDbusProperty<uint64_t>(
+    return pldm::utils::DBusHandler().getDbusProperty<uint8_t>(
         dBusMap->objectPath.c_str(), dBusMap->propertyName.c_str(),
         dBusMap->interface.c_str());
 }
